Generalize barrel rendering code
