+++ src/main/resources/assets/stockpile/blockstates/potion_barrel.json
@@ -1,0 +1,27 @@
+{
+  "variants": {
+    "facing=down": {
+      "model": "stockpile:block/barrel",
+      "x": 90
+    },
+    "facing=up": {
+      "model": "stockpile:block/barrel",
+      "x": 270
+    },
+    "facing=north": {
+      "model": "stockpile:block/barrel"
+    },
+    "facing=south": {
+      "model": "stockpile:block/barrel",
+      "y": 180
+    },
+    "facing=west": {
+      "model": "stockpile:block/barrel",
+      "y": 270
+    },
+    "facing=east": {
+      "model": "stockpile:block/barrel",
+      "y": 90
+    }
+  }
+}
